Inventory & Health UI system

+++ src/Components/HealthComponent.h
@@ -1,12 +1,10 @@
 #ifndef HEALTHCOMPONENT_H
 #define HEALTHCOMPONENT_H
 
-#include <glm/glm.hpp>
-
 struct HealthComponent
 {
-    float healthAmount;
-    HealthComponent(float amt = 1) : healthAmount(amt) {}
+    int healthAmount;
+    HealthComponent(int amt = 1) : healthAmount(amt) {}
 };
 
 #endif
